requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-openai
langchain_google_genai
google-genai
langchain_anthropic
langchain-xai
langchain-qdrant==0.2.0
langchain-postgres==0.0.15
qdrant-client==1.12.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
streamlit
fastapi
uvicorn
gunicorn
python-multipart
psycopg2-binary
nest-asyncio
psycopg[binary]
fastembed
httpx
aiofiles